requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
llama-index-core==0.9.37
llama-index-embeddings-huggingface==0.1.5
pydantic==2.5.0
python-dotenv==1.0.0
